feat: add week view with some options
It add a week view that reduce the view of the calendar only to a number specified of weeks, by default 1. It add a new even fired when the weeks change. The behaviour in change of the week depends of the continuous configuration. If continuous is active the calendar will change the weeks without any checks of months. If not in the change will be check if first day was shown and also if the weeks was filled with days of the current month.